% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateItems.R
\name{GenerateItems}
\alias{GenerateItems}
\title{Generate Symbol Search Items}
\usage{
GenerateItems(items = 20)
}
\arguments{
\item{items}{Numeric value of at least 3. The number of symbol search items
that will be generated, the default is 20.}
}
\value{
A list of variables of length 7, of which for half one of the first
two numbers is repeated in the next 5 numbers, and for the other
half the 7 numbers are all different. In addition, a answer variable
is generated with the correct answers to each item.
}
\description{
A function that generates an editable amount of items expressed in
numbers that can later be used to translate to symbols for the symbol
search sub-test.
}
\examples{
 my_items <- GenerateItems(items = 10)
 # Use this in a a shiny app
 ui <- fluidPage(
 # add the question to the top
 mainPanel(
 "Is either one of the two symbols on the left, also on the right side?",
 width = 20),
 # the first of seven buttons that all have one of the symbols of the first item
 actionButton(
 inputId = "symbol1",
 all_symbols[my_items[["items"]][["item_1"]][1]],
 class = "btn-lg btn-primary",
 disabled="disabled"))

}
