\name{GenerateItems}
\alias{GenerateItems}
\title{Generate Items for the Symbol Search}
\usage{
GenerateItems(items)
}
\description{
Generates a self-chosen amount of items for the symbol search app, and records
  the answers to the items in a seperate list.
}
\examples{
GenerateItems(items = 10)
}
